pyproject.toml
[tool.poetry]
name = "same-cli"
version = "0.1.0"
description = ""
authors = ["David Aronchick <aronchick@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.*"
pre-commit = "^2.9.3"
pathlib2 = "^2.3.5"
numpy = "^1.19.3"
jupyter = "^1.0.0"
kfp = "^1.4.0"
dill = "^0.3.3"
tensorflow = "^2.5.0"
plotly = "^4.14.3"
matplotlib = "^3.4.2"
pandas = "^1.2.4"
chart-studio = "^1.1.0"
azureml-core = {version = "^1.29.0", python = ">=3.8,<3.9"}
azureml-pipeline = {version = "^1.29.0", python = ">=3.8,<3.9p"}
scipy = "^1.6.3"
python-dotenv = "^0.17.1"
pipreqs = "^0.4.10"

[tool.poetry.dev-dependencies]
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
